🎀 [CHORE] : postID 페이지, subHeader 컴포넌트 공통 API 합치기
