+++ src/components/SubHeader/SubHeader.jsx
@@ -3,7 +3,7 @@ import * as S from './SubHeader.style';
 import AllEmoji from '../../assets/icon/arrow_down.svg';
 import AddEmoji from '../../assets/icon/add-24.svg';
 import Share from '../../assets/icon/share-24.svg';
-import { getRecipientData, getEmojiData, postEmoji } from './api';
+import { getEmojiData, postEmoji } from './api';
 import Emoji from '../PaperListEmojiBadge/PaperListEmojiBadge';
 import EmojiModal from '../subHeaderModal/showImgModal/EmojiModal';
 import EmojiPicker from 'emoji-picker-react';
@@ -11,48 +11,12 @@ import KakaoModal from '../subHeaderModal/SubHeaderKaKao/KakaoModal';
 import PaperListFromBadge from '../PaperListFromBadge';
 
 const SubHeader = ({ value }) => {
-  const [profileData, setProfileData] = useState([]);
   const [emojiData, setEmojiData] = useState([]);
   const [modalOpen, setModalOpen] = useState(false);
   const [pickerOpen, setPickerOpen] = useState(false);
   const [kakaoOpen, setKakaoOpen] = useState(false);
   const [resultPostEmoji, setResultPostEmoji] = useState(null);
-  const [userData, setUserData] = useState({
-    name: null,
-    messageCount: 0,
-    reactionCount: 0,
-    topReactions: [],
-  });
-
   const pickerRef = useRef(null);
-
-  const getUserData = async (userID) => {
-    const {
-      name,
-      messageCount,
-      recentMessages,
-      reactionCount,
-      topReactions,
-      error,
-    } = await getRecipientData(userID);
-    if (error) {
-      return;
-    }
-
-    setUserData({
-      name,
-      messageCount,
-      reactionCount,
-      topReactions,
-    });
-
-    const url = [
-      recentMessages?.[0]?.profileImageURL,
-      recentMessages?.[1]?.profileImageURL,
-      recentMessages?.[2]?.profileImageURL,
-    ];
-    setProfileData(url);
-  };
 
   const getAllEmojiData = async (userID) => {
     const { results, error } = await getEmojiData(userID);
@@ -89,7 +53,6 @@ const SubHeader = ({ value }) => {
   };
 
   useEffect(() => {
-    getUserData(value.userID);
     getAllEmojiData(value.userID);
 
     const handleClickOutside = (event) => {
@@ -107,17 +70,17 @@ const SubHeader = ({ value }) => {
   return (
     <S.SubHeader>
       <S.HeaderContent>
-        <S.UserName>To.{userData.name}</S.UserName>
+        <S.UserName>To.{value.userData.name}</S.UserName>
         <S.UserInfo>
           <S.PaperCnt>
             <S.ProfileCnt>
               <PaperListFromBadge
-                imgUrls={profileData}
-                count={userData.messageCount}
+                imgUrls={value.profileData}
+                count={value.userData.messageCount}
               />
             </S.ProfileCnt>
             <S.CntText>
-              <S.Strong>{userData.messageCount}</S.Strong> 개의 카드가
+              <S.Strong>{value.userData.messageCount}</S.Strong> 개의 카드가
               도착했어요!
             </S.CntText>
           </S.PaperCnt>
